+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to
 - Add edX course generator
 - Add edX enrollment event tests
 - Add edX problem_check event tests
+- Add edX video event tests
 
 
 [Unreleased]: https://github.com/openfun/learning-analytics-playground/commits/main